Buceo de Focas Monje (Monk Seals)

Data del artículo open access "Integrating multiple technologies to understand the foraging behaviour of Hawaiian monk seals

data(monkseals)

Format

An object of class "cross"; see read.cross.

Source

Dryad digital repository

References

Wilson, K., Littnan, C., Halpin, P., & Read, A. (2017). Integrating multiple technologies to understand the foraging behaviour of Hawaiian monk seals. Royal Society open science, 4(3), 160703 (Website)

Examples

data(monkseals)
mean(monkseals$odbaProm, na.omit = TRUE)
#> [1] 0.1541341